Added option to build for iOS simulator only.
git-svn-id: https://emgucv.svn.sourceforge.net/svnroot/emgucv/trunk@1786 d7f09016-e345-0410-b530-edf29a71df78

+++ ios/build.sh
@@ -1,26 +1,28 @@
 cd ..
 
-rm -f CMakeCache.txt
-ios/configure-device_xcode.sh
-rm -rf ios/armv6 lib/Release bin/Release
-xcodebuild -sdk iphoneos -configuration Release ARCHS="armv6" -target ALL_BUILD clean build
-mkdir -p ios/armv6 
-cp -r lib/Release/* ios/armv6/
-cp -r bin/Release/* ios/armv6/
-cd ios/armv6
-libtool -static -o libemgucv_armv6.a *.a
-cd ../..
-
-rm -f CMakeCache.txt
-ios/configure-device_xcode.sh
-rm -rf ios/armv7 lib/Release bin/Release
-xcodebuild -sdk iphoneos -configuration Release ARCHS="armv7" -target ALL_BUILD clean build
-mkdir -p ios/armv7 
-cp -r lib/Release/* ios/armv7/
-cp -r bin/Release/* ios/armv7/
-cd ios/armv7
-libtool -static -o libemgucv_armv7.a *.a
-cd ../..
+if [ "$1" != "simulator" ]; then    
+    rm -f CMakeCache.txt
+    ios/configure-device_xcode.sh
+    rm -rf ios/armv6 lib/Release bin/Release
+    xcodebuild -sdk iphoneos -configuration Release ARCHS="armv6" -target ALL_BUILD clean build
+    mkdir -p ios/armv6 
+    cp -r lib/Release/* ios/armv6/
+    cp -r bin/Release/* ios/armv6/
+    cd ios/armv6
+    libtool -static -o libemgucv_armv6.a *.a
+    cd ../..
+    
+    rm -f CMakeCache.txt
+    ios/configure-device_xcode.sh
+    rm -rf ios/armv7 lib/Release bin/Release
+    xcodebuild -sdk iphoneos -configuration Release ARCHS="armv7" -target ALL_BUILD clean build
+    mkdir -p ios/armv7 
+    cp -r lib/Release/* ios/armv7/
+    cp -r bin/Release/* ios/armv7/
+    cd ios/armv7
+    libtool -static -o libemgucv_armv7.a *.a
+    cd ../..
+fi
 
 rm -f CMakeCache.txt
 ios/configure-simulator_xcode.sh
@@ -35,7 +37,12 @@ cd ../..
 
 rm -rf ios/universal
 mkdir -p ios/universal
-lipo -create -output ios/universal/libemgucv.a ios/armv6/libemgucv_armv6.a ios/armv7/libemgucv_armv7.a ios/i386/libemgucv_i386.a
+if [ "$1" == "simulator" ]; then
+    cp ios/i386/libemgucv_i386.a ios/universal/libemgucv.a
+else
+    lipo -create -output ios/universal/libemgucv.a ios/armv6/libemgucv_armv6.a ios/armv7/libemgucv_armv7.a ios/i386/libemgucv_i386.a
+fi
+
 mkdir -p Emgu.CV/PInvoke/MonoTouch
 cp -f ios/universal/libemgucv.a Emgu.CV/PInvoke/MonoTouch
